Refactor: organise files into packages within ui
