+++ src/main/java/powerbake/address/ui/persontab/PersonCard.java
@@ -1,4 +1,4 @@
-package powerbake.address.ui;
+package powerbake.address.ui.persontab;
 
 import java.util.Comparator;
 
@@ -8,6 +8,7 @@ import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
 import powerbake.address.model.person.Person;
+import powerbake.address.ui.UiPart;
 
 /**
  * An UI component that displays information of a {@code Person}.
